\name{GOGANPA-package}
\alias{GOGANPA-package}
\docType{package}
\title{
GO-Functional-Network-based Gene-Set-Analysis
}
\description{
Accounting for genes' functional-non-equivalence within pathways in classical Gene-set-analysis.
}
\details{
\tabular{ll}{
Package: \tab GOGANPA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Depends: \tab GANPA, WGCNA\cr
Suggests: \tab GANPA, WGCNA, GANPAdata\cr
Date: \tab 2011-12-30\cr
License: \tab GPL-2\cr
}

}
\author{
Billy Chang

Maintainer: Billy Chang <billy.chang@utoronto.ca>
}
\references{
Chang, B, Kustra, R, and Tian, WD, (2012) Functional-Network-Based Gene-Set-Analysis using Gene Ontology. Submitted.

Fang, ZY, Tian, WD, and Ji, HB. (September 6, 2011) A network-based gene-weighting approach for pathway analysis. Cell Research. Advanced Publication.
}
